feat(consensus): consensus update

+++ consensus/src/main/java/io/nuls/consensus/service/intf/BlockService.java
@@ -12,11 +12,11 @@ public interface BlockService {
 
     Block getGengsisBlockFromDb();
 
-    int getLocalHeight();
+    long getLocalHeight();
 
     byte[] getLocalHighestHash();
 
-    int getBestHeight();
+    long getBestHeight();
 
     byte[] getBestHash();
 
@@ -26,7 +26,7 @@ public interface BlockService {
 
     Block getBlockByHash(String hash);
 
-    Block getBlockByHeight(int height);
+    Block getBlockByHeight(long height);
 
     void save(Block block);
